added some pretty pages that don't do anything
Still couldn't figure out how to make a table that isn't in the main class, even though I spent over 3 hours testing and researching. For some reason, if it is put in the main class, it's actually a breeze to generate. Otherwise, it's just hell. So if you find a way to pull something from the main class into another window, just let me know and I can hook you up with some working code. I've also included a "Search Auction" page...that doesn't work. Since I couldn't even figure out how to write plain text from a list into a box, I couldn't actually do anything with search results. Sorry about that. I do have ONE page that works. "Create Auction" does generate an auction when the button is pushed...but it still has issues as it was written at about 4am. The description box is stuck in 'single line mode' or something and despite it creating and writing an auction when the button is pressed, there is no indication that you actually accomplished anything.